Change model invert func location
