+++ curiefense/curieconf/server/curieconf/confserver/v1/api.py
@@ -357,6 +357,46 @@ schema_type_map = {
     "wafrules": content_filter_rule_schema,
 }
 
+# Create new models inverting field name and its attribute
+
+def _field_invert_names(field):
+    """
+    Helper function to recurse over child fields incase of Nested/Wildcard/List fields.
+
+    Args:
+        field (fields.Raw): field to recurse. Being mutated.
+
+    Returns
+        fields.Raw: converted field
+    """
+
+    if isinstance(field, fields.Nested):
+        field.model = model_invert_names(field.model)
+    elif isinstance(field, fields.List) or isinstance(field, fields.Wildcard):
+        field.container = _field_invert_names(field.container)
+    return field
+
+def model_invert_names(model):
+    """
+    Invert key names in a model using fields attribute if exists.
+
+    Args:
+        model (Model): model to invert.
+
+    Returns
+        Model: inverted model
+    """
+
+    mod = model.clone(model.name)
+    for key in list(mod):
+        _field_invert_names(mod[key])
+        if mod[key].attribute:
+            new_key = mod[key].attribute
+            mod[new_key] = mod[key]
+            mod[new_key].attribute = key
+            del mod[key]
+    return mod
+
 ################
 ### CONFIGS ###
 ################
@@ -531,7 +571,7 @@ class DocumentResource(Resource):
         "Create a new complete document"
         if document not in models:
             abort(404, "document does not exist")
-        data = marshal(request.json, utils.model_invert_names(models[document]), skip_none=True)
+        data = marshal(request.json, model_invert_names(models[document]), skip_none=True)
         res = current_app.backend.documents_create(
             config, utils.vconvert(document, "v1"), data
         )
@@ -541,7 +581,7 @@ class DocumentResource(Resource):
         "Update an existing document"
         if document not in models:
             abort(404, "document does not exist")
-        data = marshal(request.json, utils.model_invert_names(models[document]), skip_none=True)
+        data = marshal(request.json, model_invert_names(models[document]), skip_none=True)
         res = current_app.backend.documents_update(
             config, utils.vconvert(document, "v1"), data
         )
@@ -608,7 +648,7 @@ class EntriesResource(Resource):
         "Create an entry in a document"
         if document not in models:
             abort(404, "document does not exist")
-        data = marshal(request.json, utils.model_invert_names(models[document]), skip_none=True)
+        data = marshal(request.json, model_invert_names(models[document]), skip_none=True)
         res = current_app.backend.entries_create(
             config, utils.vconvert(document, "v1"), data
         )
@@ -632,7 +672,7 @@ class EntryResource(Resource):
             abort(404, "document does not exist")
         isValid = validateJson(request.json, document)
         if isValid:
-            data = marshal(request.json, utils.model_invert_names(models[document]), skip_none=True)
+            data = marshal(request.json, model_invert_names(models[document]), skip_none=True)
             res = current_app.backend.entries_update(
                 config, utils.vconvert(document, "v1"), entry, data
             )
