+++ curiefense/curieconf/test/test_utils.py
@@ -3,7 +3,6 @@ from curieconf import utils
 import json
 import codecs
 import base64
-from flask_restplus import fields, model
 
 binvec_hex = (
     "b70a1da09a4998bd56b083d76bf528053c9b924bbb07168792151a5a177bbaa232949a8600bcb2"
@@ -68,46 +67,6 @@ def test_bytes2jblob_json():
     }
     res2 = utils.jblob2bytes(res)
     assert res2 == vec
-
-def test_model_invert_names():
-    mod1 = model.Model("test", {
-            "test":fields.String(attribute="test2")
-    })
-    res = utils.model_invert_names(mod1)
-    assert res.name == mod1.name and type(res['test2']) is fields.String \
-        and res['test2'].attribute == 'test'
-
-    mod2 = model.Model("test", {
-            "test":fields.Nested(mod1, attribute="test2")
-    })
-    res = utils.model_invert_names(mod2)
-    assert res.name == mod2.name \
-        and type(res['test2']) is fields.Nested \
-        and res['test2'].attribute == 'test' \
-        and type(res['test2'].model['test2']) is fields.String \
-        and res['test2'].model['test2'].attribute == 'test'
-
-    mod3 = model.Model("test", {
-            "test":fields.List(mod1, attribute="test2")
-    })
-    res = utils.model_invert_names(mod3)
-    assert res.name == mod3.name \
-        and type(res['test2']) is fields.List \
-        and res['test2'].attribute == 'test' \
-        and type(res['test2'].container['test2']) is fields.String \
-        and res['test2'].container['test2'].attribute == 'test'
-
-    mod4 = model.Model("test", {
-            "test*":fields.Wildcard(mod2, attribute="test2*")
-    })
-    res = utils.model_invert_names(mod4)
-    assert res.name == mod4.name \
-        and type(res['test2']) is fields.Wildcard \
-        and res['test2'].attribute == 'test' \
-        and type(res['test2'].container['test2']) is fields.Nested \
-        and res['test2'].container['test2'].attribute == 'test' \
-        and type(res['test2'].container['test2'].model['test2']) is fields.String \
-        and res['test2'].container['test2'].model['test2'].attribute == 'test'
 
 def test_vconvert():
     assert utils.vconvert("urlmaps", "v1", False) == "securitypolicies"
